Finished up to 4.22.

+++ controllers/blogs.js
@@ -11,11 +11,7 @@ blogsRouter.get("/", async (req, res) => {
 
 blogsRouter.post("/", async (req, res, next) => {
   const { title, author, url, likes } = req.body;
-  const decodedToken = jwt.verify(req.token, config.SECRET);
-  if (decodedToken.id === null) {
-    return res.status(401).json({ error: "token invalid" });
-  }
-  const user = await User.findById(decodedToken.id);
+  const user = req.user;
   const blog = new Blog({
     title: title,
     author: author,
@@ -37,17 +33,30 @@ blogsRouter.post("/", async (req, res, next) => {
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
+  const { id } = req.body;
+  const user = req.user;
+  const blog = await Blog.findById(id);
+  if (blog.user.toString() !== user.id.toString()) {
+    return res.status(400).json({ error: "not authorized" });
+  }
   const result = await Blog.deleteOne(req.body);
   res.status(200).json(result);
 });
 
 blogsRouter.put("/:id", async (req, res) => {
+  const { title, author, url, likes } = req.body;
+  const user = req.user;
   let oldBlog = await Blog.findById(req.body.id);
+
+  if (oldBlog.user.toString() !== user.id.toString()) {
+    return res.status(400).json({ error: "not authorized" });
+  }
+
   oldBlog = {
-    title: req.body.title,
-    author: req.body.author,
-    url: req.body.url,
-    likes: req.body.likes,
+    title: title,
+    author: author,
+    url: url,
+    likes: likes,
   };
   const result = await Blog.updateOne(oldBlog);
   res.status(200).json(result);
